Refactor out the ChromeAttachment, replacing it with a thorough and deep rewrite. Honest.
--- also added license
